[IMP] update display string formate to hour:min, make it user specific so every user will get his own fetchmail timing
