+++ mail_check_immediately/__openerp__.py
@@ -11,7 +11,7 @@
     'currency': 'EUR',
     'depends': ['base', 'web', 'fetchmail', 'mail'],
     'data': [
-        'views.xml',
+        'views/views.xml',
     ],
     'qweb': [
         "static/src/xml/main.xml",
