+++ mail_check_immediately/models/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
-from . import models
+
+import models
