+++ mail_check_immediately/static/src/js/main.js
